@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.16.1 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,13 +27,13 @@ dependencies:
 
 
   cupertino_icons: ^1.0.0
-  firebase_core: ^0.5.3
-  firebase_auth: ^0.18.4+1
-  modal_progress_hud: ^0.1.3
+  firebase_core: ^1.21.1
+  firebase_auth: ^3.7.0
   flutter_svg: ^1.1.0
   provider: ^6.0.3
   hexcolor: ^2.0.7
   flutter_spinkit: ^5.1.0
+  cloud_firestore: ^3.4.6
 
 dev_dependencies:
   flutter_test:
