@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:modal_progress_hud/modal_progress_hud.dart';
 import '../../constants.dart';
 import '../Login/login_screen_top_image.dart';
 
@@ -11,16 +10,14 @@ class RegistrationScreen extends StatefulWidget {
 
 class _RegistrationScreenState extends State<RegistrationScreen> {
   final _auth = FirebaseAuth.instance;
-  String email;
-  String password;
+  late String email;
+  late String password;
   bool showSpinner = false;
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         backgroundColor: Colors.white,
-        body: ModalProgressHUD(
-        inAsyncCall: showSpinner,
-        child: SingleChildScrollView(
+        body: SingleChildScrollView(
         child: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: <Widget>[
@@ -83,9 +80,7 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
                   "Sign Up".toUpperCase(),
                 ),
                 onPressed: () async {
-                  setState(() {
-                    showSpinner = true;
-                  });
+                  buildShowDialog(context);
                   try {
                     final newUser = await _auth.createUserWithEmailAndPassword(
                         email: email, password: password);
@@ -95,9 +90,7 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
                   } catch (e) {
                     print(e);
                   }
-                  setState(() {
-                    showSpinner = false;
-                  });
+
                 }),
             ),
             ],
@@ -112,7 +105,16 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
         ),
         ]),
         ),
-        ),
     );
+  }
+  buildShowDialog(BuildContext context) {
+    return showDialog(
+        context: context,
+        barrierDismissible: false,
+        builder: (BuildContext context) {
+          return Center(
+            child: CircularProgressIndicator(),
+          );
+        });
   }
 }
